Move the multiple choice element in to /shared
